+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - `Innmind\HttpTransport\Curl::maxConcurrency()` use `::map()` instead
 - `Innmind\HttpTransport\Curl::disableSSLVerification()` use `::map()` instead
 - `Innmind\HttpTransport\Curl::proxy()` use `::map()` instead
+- `Innmind\HttpTransport\Curl::heartbeat()` (as it was intended for internal use)
 
 ## 8.1.0 - 2025-09-18